portable-rag/requirements.txt
#fastapi
#uvicorn
ollama
python-dotenv
faiss-cpu
numpy

streamlit
reportlab
